(trunk) Compile fixes for Ubuntu 17.10.
git-svn-id: http://svn.gplates.org/projects/gplates/trunk@18546 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c

+++ src/qt-widgets/ScalarField3DDepthLayersPage.cc
@@ -25,8 +25,10 @@
 
 #include <algorithm>
 #include <cmath>
+#include <cstddef> // For std::size_t
 #include <boost/foreach.hpp>
 #include <boost/bind.hpp>
+#include <boost/cast.hpp>
 #include <boost/weak_ptr.hpp>
 #include <QTableWidgetItem>
 #include <QItemDelegate>
@@ -801,7 +803,7 @@ GPlatesQtWidgets::ScalarField3DDepthLayersPage::add_files_to_sequence(
 	std::vector< boost::optional<double> > depths;
 	deduce_depths(depths, file_infos);
 	GPlatesGlobal::Assert<GPlatesGlobal::AssertionFailureException>(
-			depths.size() == file_infos.size(),
+			depths.size() == boost::numeric_cast<std::size_t>(file_infos.size()),
 			GPLATES_ASSERTION_SOURCE);
 
 	ScalarField3DDepthLayersSequence new_sequence;
